remove ThreadStatic and small cleanups
